+++ core/samples/logging/main.cpp
@@ -11,10 +11,6 @@
 #include <cubos/core/reflection/external/unordered_map.hpp>
 /// [External reflection includes]
 
-struct TypeWithoutReflection
-{
-};
-
 int main()
 {
     /// [Set logging level]
@@ -42,6 +38,5 @@ int main()
     CUBOS_INFO("An integer: {}", 1);
     CUBOS_INFO("A glm::vec3: {}", glm::vec3(0.0F, 1.0F, 2.0F));
     CUBOS_INFO("An std::unordered_map: {}", std::unordered_map<int, const char*>{{1, "one"}, {2, "two"}});
-    CUBOS_INFO("A type without reflection: {}", TypeWithoutReflection{});
 }
 /// [Logging macros with arguments]