+++ engine/src/render/g_buffer_rasterizer/plugin.cpp
@@ -48,6 +48,8 @@ namespace
 
     struct State
     {
+        CUBOS_ANONYMOUS_REFLECT(State);
+
         ShaderPipeline pipeline;
         ShaderBindingPoint perSceneBP;
         ShaderBindingPoint perMeshBP;