+++ engine/src/render/shadow_atlas_rasterizer/plugin.cpp
@@ -41,6 +41,8 @@ namespace
 
     struct State
     {
+        CUBOS_ANONYMOUS_REFLECT(State);
+
         ShaderPipeline pipeline;
         ShaderBindingPoint perSceneBP;
         ShaderBindingPoint perMeshBP;