+++ engine/src/render/shadows/directional_caster.cpp
@@ -1,6 +1,7 @@
 #include <cubos/core/ecs/reflection.hpp>
 #include <cubos/core/reflection/external/glm.hpp>
 #include <cubos/core/reflection/external/primitives.hpp>
+#include <cubos/core/reflection/external/vector.hpp>
 
 #include <cubos/engine/render/shadows/directional_caster.hpp>